+++ harmony-homes/src/components/organisms/howitwork.tsx
